backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.25
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
paho-mqtt==1.6.1